arr-setup/pyproject.toml
[tool.poetry]
name = "arr_setup"
version = "0.1.0"
description = ""
authors = ["barsa2000 <barsanti.filippo2@gmail.com>"]
license = "MIT"
readme = "README.md"
# packages = [
#     {include = "arr_setup"}
# ]

[tool.poetry.dependencies]
python = "^3.12"
pyarr = "^5.2.0"
pyyaml = "^6.0.1"
typer = "^0.12.3"
loguru = "^0.7.2"
rich = "^13.7.1"

[tool.poetry.scripts]
arr_setup = "arr_setup.main:cli"


[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
black = "^24.4.2"
flake8 = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79